requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
faiss-cpu>=1.7.4
numpy>=1.21.0
requests>=2.26.0
streamlit>=1.12.0
google-generativeai>=0.3.0
tokenizers==0.18.3
transformers==4.36.0
sentence-transformers==2.2.2